refactor saveOutput
saveoutput squash
